moved game files into game directory for better organization
