+++ KittyCount.activity/game/Utilities.py
