+++ KittyCount.activity/main.py
@@ -1,9 +1,6 @@
 import pygame
-from Utilities import Utilities
-from UI import UI
-from NumberLine import NumberLine
-from character import character
-import time
+from game.Utilities import Utilities
+from game.UI import UI
 
 def main():
     pygame.init()
@@ -12,7 +9,6 @@ def main():
     #screen = pygame.display.set_mode((0, 0), pygame.FULLSCREEN)
     screen = pygame.display.set_mode((0, 0))  # windowed app for DEV only
     screen.fill((255, 255, 255))
-    pygame.display.flip()
     done = False
 
     # setup variables
@@ -22,17 +18,6 @@ def main():
     # display UI
     ui = UI(screen, w/2 - 200, h/2 + 150)
 
-    #display Number Line
-    number_line = NumberLine(screen, w, h)
-    number_line.display()
-    #display cat
-    cat_x, cat_y = number_line.circle_pos[7]
-    cat = character(cat_x, cat_y, "../Assets/cat1.png", 100, 100, screen)
-
-            
-    #display mouse
-    mouse_x, mouse_y = number_line.circle_pos[3]
-    mouse = character (mouse_x, mouse_y, "../Assets/mouse3.png", 50, 50, screen)
 
     clock = pygame.time.Clock()
     # exit logic
@@ -49,8 +34,6 @@ def main():
         # draw UI
         screen.fill((255, 255, 255))
         ui.display(events)
-        number_line.display()
-        mouse.display()
         
 
         pygame.display.update()
